lambda/requirements.txt
aws-xray-sdk==2.6.0
beautifulsoup4==4.9.3
botocore==1.20.12
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
feedparser==6.0.2
future==0.18.2
idna==2.10
jmespath==0.10.0
jsonpickle==2.0.0
python-dateutil==2.8.1
requests==2.25.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
wrapt==1.12.1
